\name{redistrict}
\alias{redistrict}

\title{Redistricting Monte-Carlo data}

\description{ Precinct-level observations for a hypothetical jurisdiction 
with four proposed districts.  
}

\usage{data(redistrict)}

\format{A table containing 150 observations and 9 variables:  
\describe{
\item{precinct}{precinct identifier}
\item{district}{proposed district number}
\item{avg.age}{average age}
\item{per.own}{percent homeowners}
\item{black}{number of black voting age persons}
\item{white}{number of white voting age persons}
\item{hispanic}{number of hispanic voting age persons}
\item{total}{total number of voting age persons}
\item{dem}{Number of votes for the Democratic candidate}
\item{rep}{Number of votes for the Republican candidate}
\item{no.vote}{Number of non voters}
}}

\source{Daniel James Greiner}

\keyword{datasets}
