\name{senc}
\alias{senc}

\title{Party registration in south-east North Carolina}

\description{ Registration data for White, Black, and Native American voters in 
eight counties of 
south-eastern North Carolina in 2001.  
}

\usage{data(senc)}

\format{A table containing 212 observations and 18 variables:  
\describe{
\item{county}{county name}
\item{precinct}{precinct name}
\item{total}{number of registered voters in precinct}
\item{white}{number of White registered voters}
\item{black}{number of Black registered voters}
\item{natam}{number of Native American registered voters}
\item{dem}{number of registered Democrats}
\item{rep}{number of registered Republicans}
\item{other}{number of registered voters without major party affiliation}
\item{whdem}{number of White registered Democrats}
\item{whrep}{number of White registered Republicans}
\item{whoth}{number of White registered voters without major party affiliation}
\item{bldem}{number of Black registered Democrats}
\item{blrep}{number of Black registered Republicans}
\item{bloth}{number of Black registered voters without major party affiliation}
\item{natamdem}{number of Native American registered Democrats}
\item{natamrep}{number of Native American registered Republicans}
\item{natamoth}{number of Native American registered voters without major party 
affiliation}
}}

\source{Excerpted from North Carolina General Assembly 2001 redistricting data, 
\url{https://www2.ncleg.net/RnR/Redistricting/BaseData2001}}

\keyword{datasets}
